+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    preset: "ts-jest",
+    testEnvironment: "node",
+    testMatch: ["**/**/*.test.ts"],
+    verbose: true,
+    forceExit: false,
+  };
